+++ __tests__/html/timestamp.changeSendTimeout.js
@@ -1,0 +1,7 @@
+/**
+ * @jest-environment ./__tests__/html/__jest__/WebChatEnvironment.js
+ */
+
+describe('timestamp', () => {
+  test('change send timeout on-the-fly', () => runHTMLTest('timestamp.changeSendTimeout.html'));
+});