+++ packages/testharness/src/token/index.js
@@ -1,5 +1,11 @@
+import fetchDirectLineAppServiceExtensionToken from './fetchDirectLineAppServiceExtensionToken';
 import fetchDirectLineSpeechCredentials from './fetchDirectLineSpeechCredentials';
 import fetchDirectLineToken from './fetchDirectLineToken';
 import fetchSpeechServicesCredentials from './fetchSpeechServicesCredentials';
 
-export { fetchDirectLineSpeechCredentials, fetchDirectLineToken, fetchSpeechServicesCredentials };
+export {
+  fetchDirectLineAppServiceExtensionToken,
+  fetchDirectLineSpeechCredentials,
+  fetchDirectLineToken,
+  fetchSpeechServicesCredentials
+};